# backend/app/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IntelliDesk AI — API Specification",
    "description": "Comprehensive interactive REST API documentation for the IntelliDesk AI platform, including user authentication, ticketing, SLA management, incident handling, and AI-powered knowledge base operations.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Local Gateway (default)"
    },
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Flask Backend Direct"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/tickets": {
      "get": {
        "tags": ["Tickets"],
        "summary": "List Tickets",
        "description": "Retrieve tickets. Scoped automatically by role (employees see own, agents/managers see all scoped to department).",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["new", "open", "pending", "resolved", "closed"] } },
          { "name": "priority", "in": "query", "schema": { "type": "string", "enum": ["low", "medium", "high", "critical"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of tickets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Ticket" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tickets"],
        "summary": "Create Ticket",
        "description": "Submit a new support ticket.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "description", "category"],
                "properties": {
                  "title": { "type": "string", "example": "Printer setup issues on 3rd floor" },
                  "description": { "type": "string", "example": "Cannot connect to the office printer from my laptop." },
                  "category": { "type": "string", "example": "Hardware" },
                  "priority": { "type": "string", "enum": ["low", "medium", "high", "critical"], "default": "medium" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "data": { "$ref": "#/components/schemas/Ticket" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{id}": {
      "get": {
        "tags": ["Tickets"],
        "summary": "Get Ticket Details",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Ticket details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "data": { "$ref": "#/components/schemas/Ticket" }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Tickets"],
        "summary": "Update Ticket Status / Priority",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "enum": ["open", "pending", "resolved", "closed"] },
                  "priority": { "type": "string", "enum": ["low", "medium", "high", "critical"] },
                  "resolution_notes": { "type": "string", "example": "Configured the printer server IP manually." }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Ticket updated successfully" }
        }
      },
      "delete": {
        "tags": ["Tickets"],
        "summary": "Soft Delete Ticket",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "Ticket soft-deleted" }
        }
      }
    },
    "/tickets/{id}/assign": {
      "put": {
        "tags": ["Tickets"],
        "summary": "Assign Ticket to Agent",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["assigned_to_id"],
                "properties": {
                  "assigned_to_id": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Ticket assigned successfully" }
        }
      }
    },
    "/tickets/bulk-update": {
      "put": {
        "tags": ["Tickets"],
        "summary": "Bulk Update Tickets",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ticket_ids"],
                "properties": {
                  "ticket_ids": { "type": "array", "items": { "type": "integer" } },
                  "status": { "type": "string" },
                  "assigned_to_id": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Tickets updated successfully" }
        }
      }
    },
    "/tickets/{id}/comments": {
      "get": {
        "tags": ["Tickets"],
        "summary": "List Ticket Comments",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "List of comment objects" }
        }
      },
      "post": {
        "tags": ["Tickets"],
        "summary": "Add Comment to Ticket",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["body"],
                "properties": {
                  "body": { "type": "string", "example": "Is there any update on this?" },
                  "is_private": { "type": "boolean", "default": false }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Comment added" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Pass your JWT token in the Authorization header: `Bearer <token>`"
      }
    },
    "schemas": {
...
      "Ticket": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "ticket_number": { "type": "string", "example": "TKT-2026-0001" },
          "title": { "type": "string", "example": "VPN connection issues" },
          "description": { "type": "string" },
          "status": { "type": "string", "example": "open" },
          "priority": { "type": "string", "example": "medium" },
          "category": { "type": "string", "example": "Network" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}